+++ tests/core/test_mythic_lockdown.lua
@@ -193,7 +193,7 @@ return function()
   -- -----------------------------------------------------------------------
   do
     local Constants = require("WhisperMessenger.Core.Constants")
-    assert(Constants.VERSION == "v1.1.3", "VERSION should be v1.1.3, got: " .. tostring(Constants.VERSION))
+    assert(Constants.VERSION == "v1.1.4", "VERSION should be v1.1.4, got: " .. tostring(Constants.VERSION))
   end
 
   -- -----------------------------------------------------------------------